oxirgi o'zgarishlar qoshildi
